+++ app/api/rpcApi.js
@@ -170,6 +170,25 @@ function getBlockByHeight(blockHeight) {
 	});
 }
 
+function getBlockHeaderByHash(blockhash) {
+	return getRpcDataWithParams({method:"getblockheader", parameters:[blockhash]});
+}
+
+function getBlockHeaderByHeight(blockHeight) {
+	return new Promise(function(resolve, reject) {
+		getRpcDataWithParams({method:"getblockhash", parameters:[blockHeight]}).then(function(blockhash) {
+			getBlockHeaderByHash(blockhash).then(function(blockHeader) {
+				resolve(blockHeader);
+
+			}).catch(function(err) {
+				reject(err);
+			});
+		}).catch(function(err) {
+			reject(err);
+		});
+	});
+}
+
 function getBlockByHash(blockHash) {
 	debugLog("getBlockByHash: %s", blockHash);
 
@@ -439,6 +458,8 @@ module.exports = {
 	getNetworkHashrate: getNetworkHashrate,
 	getBlockStats: getBlockStats,
 	getBlockStatsByHeight: getBlockStatsByHeight,
+	getBlockHeaderByHash: getBlockHeaderByHash,
+	getBlockHeaderByHeight: getBlockHeaderByHeight,
 
 	minRpcVersions: minRpcVersions
 };