+++ app/coins/btc.js
@@ -64,6 +64,7 @@ module.exports = {
 	],
 	maxBlockWeight: 4000000,
 	maxBlockSize: 1000000,
+	difficultyAdjustmentBlockCount: 2016,
 	maxSupply: new Decimal(20999817.31308491), // ref: https://bitcoin.stackexchange.com/a/38998
 	targetBlockTimeSeconds: 600,
 	currencyUnits:currencyUnits,