+++ app/config.js
@@ -166,7 +166,7 @@ module.exports = {
 			]
 		},
 		subHeaderToolsList:[0, 10, 9, 4, 11, 6, 7], // indexes in "siteTools" below that are shown in the site "sub menu" (visible on all pages except homepage)
-		prioritizedToolIdsList: [0, 10, 11, 9, 3, 4, 2, 5, 1, 6, 7, 8],
+		prioritizedToolIdsList: [0, 10, 11, 9, 3, 4, 2, 5, 12, 1, 6, 7, 8],
 	},
 
 	credentials: credentials,
@@ -189,6 +189,7 @@ module.exports = {
 	/* 9 */		{name:"Mining Summary", url:"/mining-summary", desc:"Summary of recent data about miners.", fontawesome:"fas fa-chart-pie"},
 	/* 10 */	{name:"Block Stats", url:"/block-stats", desc:"Summary data for blocks in configurable range.", fontawesome:"fas fa-layer-group"},
 	/* 11 */	{name:"Block Analysis", url:"/block-analysis", desc:"Summary analysis for all transactions in a block.", fontawesome:"fas fa-angle-double-down"},
+	/* 12 */	{name:"Difficulty History", url:"/difficulty-history", desc:"Graph of difficulty changes over time.", fontawesome:"fas fa-chart-line"},
 	],
 
 	donations:{