+++ routes/apiRouter.js
@@ -41,6 +41,21 @@ router.get("/blocks-by-height/:blockHeights", function(req, res, next) {
 	});
 });
 
+router.get("/block-headers-by-height/:blockHeights", function(req, res, next) {
+	var blockHeightStrs = req.params.blockHeights.split(",");
+	
+	var blockHeights = [];
+	for (var i = 0; i < blockHeightStrs.length; i++) {
+		blockHeights.push(parseInt(blockHeightStrs[i]));
+	}
+
+	coreApi.getBlockHeadersByHeight(blockHeights).then(function(result) {
+		res.json(result);
+
+		next();
+	});
+});
+
 router.get("/block-stats-by-height/:blockHeights", function(req, res, next) {
 	var blockHeightStrs = req.params.blockHeights.split(",");